+++ src/extension/Cycode.VisualStudio.Extension.Shared/Constants.cs
@@ -4,7 +4,7 @@ namespace Cycode.VisualStudio.Extension.Shared;
 
 public static class Constants {
     public const string AppName = "visual_studio_extension";
-    public const string RequiredCliVersion = "3.6.0";
+    public const string RequiredCliVersion = "3.9.3";
 
     public const string CycodeDomain = "cycode.com";
 
@@ -17,14 +17,6 @@ public static class Constants {
 
     public const int PluginAutoSaveFlushInitialDelaySec = 0;
     public const int PluginAutoSaveFlushDelaySec = 5;
-
-    public const string SentryDsn =
-        "https://091cdc01001e4600a30ac02f1b82c4c5@o1026942.ingest.us.sentry.io/4507543901700096";
-
-    public const bool SentryDebug = false;
-    public const float SentrySampleRate = 1.0f;
-    public const bool SentrySendDefaultPii = false;
-    public const bool SentryAutoSessionTracking = true;
 
     public static readonly string PluginPath = GetPluginsPath();
     public static readonly string DefaultCliPath = GetDefaultCliPath();